refactor(test): remove 'undoableOp' from 'TestClass'
in order to split out `TestClass` to libanki:testutils
